apps/agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "prosody-agent"
version = "0.1.0"
description = "Prosody FastAPI agent scaffold"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.32,<1.0",
  "pydantic>=2.9,<3.0",
  "httpx>=0.28,<1.0",
  "openai>=2.0,<3.0",
  "pipecat-ai[webrtc,deepgram,openai,elevenlabs]>=0.0.108,<0.1",
  "pytest>=8.3,<9.0"
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
